+++ resources/js/App.vue
@@ -100,10 +100,10 @@
                     </li>
                     <li class="nav-item dropdown" v-if="loggedIn">
                         <a href="#" class="nav-link dropdown-toggle" id="user-dropdown" data-toggle="dropdown" role="button" aria-expanded="false" aria-haspopup="true">
-                            <i class="fas fa-fw fa-user"></i> {{ $auth.user().name }}
+                            <i class="fas fa-fw fa-user"></i> {{ authUser.name }}
                         </a>
                         <ul class="dropdown-menu dropdown-menu-right" aria-labelledby="user-dropdown">
-                            <router-link :to="{name: 'userpreferences', params: { id: $auth.user().id }}" class="dropdown-item">
+                            <router-link :to="{name: 'userpreferences', params: { id: authUser.id }}" class="dropdown-item">
                                 <i class="fas fa-fw fa-cog"></i> {{ $t('global.user.settings') }}
                             </router-link>
                             <a class="dropdown-item" href="#"
@@ -237,8 +237,13 @@
             }
         },
         computed: {
-            loggedIn: function() {
+            loggedIn() {
                 return this.$auth.check();
+            },
+            authUser() {
+                if(!this.loggedIn) return {};
+
+                return this.$auth.user() ? this.$auth.user() : {};
             }
         }
     }